Fix TagTest location
